parsers bluehawk member field, bluehawksource type, range/location refactor
