+++ src/bluehawk.ts
@@ -1,21 +1,51 @@
 import { makeBlockCommentTokens } from "./lexer/makeBlockCommentTokens";
 import { makeLineCommentToken } from "./lexer/makeLineCommentToken";
-import { BluehawkError, makeCstVisitor } from "./parser/makeCstVisitor";
+import { makeCstVisitor } from "./parser/makeCstVisitor";
 import { validateVisitorResult } from "./parser/validator";
 import { RootParser } from "./parser/RootParser";
+
+export interface Location {
+  line: number;
+  column: number;
+  offset: number;
+}
+
+export interface Range {
+  start: Location;
+  end: Location;
+}
+
+export interface BluehawkError {
+  message: string;
+  location: Location;
+}
 
 export class BluehawkResult {
   errors: BluehawkError[];
 }
 
+export class BluehawkSource {
+  text: string;
+  language: string;
+}
+
 export class Bluehawk {
-  run(text: string): BluehawkResult {
-    const parser = new RootParser([
-      ...makeBlockCommentTokens(/\/\*/y, /\*\//y),
-      makeLineCommentToken(/\/\//y),
-    ]);
+  parsers = new Map<string, RootParser>();
+
+  run(source: BluehawkSource): BluehawkResult {
+    if (!this.parsers.has(source.language)) {
+      this.parsers.set(
+        source.language,
+        new RootParser([
+          // TODO: map source.language to block/line comment tokens
+          ...makeBlockCommentTokens(/\/\*/y, /\*\//y),
+          makeLineCommentToken(/\/\//y),
+        ])
+      );
+    }
+    const parser = this.parsers.get(source.language);
     const { lexer } = parser;
-    const parseResult = parser.parse(text);
+    const parseResult = parser.parse(source.text);
     const visitor = makeCstVisitor(parser);
     const visitorResult = visitor.visit(parseResult.cst);
     const validateResult = validateVisitorResult(visitorResult);
